// plio-api/plio/apis/UserProfileController.ts
/* eslint-disable */

/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';
import {
    UserProfile,
    UserProfileFromJSON,
    UserProfileToJSON,
} from '../models/UserProfile';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & UserProfile;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & UserProfile;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

/**
 * no description
 */
export const UserProfileController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/plio/api/UserProfile/{id}')
    }),
    /**
     */
    create: createEndpoint<UserProfile, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/plio/api/UserProfile')
    }),
    /**
     */
    read: createEndpoint<UserProfile, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/plio/api/UserProfile/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<UserProfile>, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/UserProfile')
    }),
    /**
     */
    sparseUpdate: createEndpoint<UserProfile, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/plio/api/UserProfile/{id}')
    }),
    /**
     */
    update: createEndpoint<UserProfile, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/plio/api/UserProfile')
    }),
}