// plio-api/plio/models/UserProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { exists } from '@1023-ventures/darri-core';


export type UserProfile = {
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    changedBy?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserProfile
     */
    changedOn?: Date;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    createdBy?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserProfile
     */
    createdOn?: Date;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    concurrency?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserProfile
     */
    email?: string;
}

export const UserProfileModel = {
    id: 'id',
    tenantId: 'tenantId',
    changedBy: 'changedBy',
    changedOn: 'changedOn',
    createdBy: 'createdBy',
    createdOn: 'createdOn',
    concurrency: 'concurrency',
    name: 'name',
    userId: 'userId',
    email: 'email',
    _metadata: {
        id: { type: 'string' },
        tenantId: { type: 'string' },
        changedBy: { type: 'string' },
        changedOn: { type: 'Date' },
        createdBy: { type: 'string' },
        createdOn: { type: 'Date' },
        concurrency: { type: 'string' },
        name: { type: 'string' },
        userId: { type: 'string' },
        email: { type: 'string' },
    }
}

export function UserProfileFromJSON(json: any): UserProfile {
    return {
        'id': !exists(json, 'id') ? undefined : json['id'],
        'tenantId': !exists(json, 'tenantId') ? undefined : json['tenantId'],
        'changedBy': !exists(json, 'changedBy') ? undefined : json['changedBy'],
        'changedOn': !exists(json, 'changedOn') ? undefined : new Date(json['changedOn']),
        'createdBy': !exists(json, 'createdBy') ? undefined : json['createdBy'],
        'createdOn': !exists(json, 'createdOn') ? undefined : new Date(json['createdOn']),
        'concurrency': !exists(json, 'concurrency') ? undefined : json['concurrency'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        
    };
}

export function UserProfileToJSON(value?: UserProfile): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'id': value.id,
        'tenantId': value.tenantId,
        'changedBy': value.changedBy,
        'changedOn': value.changedOn === undefined ? undefined : value.changedOn.toISOString(),
        'createdBy': value.createdBy,
        'createdOn': value.createdOn === undefined ? undefined : value.createdOn.toISOString(),
        'concurrency': value.concurrency,
        'name': value.name,
        'userId': value.userId,
        'email': value.email,
        
    };
}